bugs

+++ phylogeny/distTree_inc_quality_phen.sh
@@ -12,9 +12,10 @@ if ($# != 5) then
 endif
 
 
+if ($2 < 1)   exit 1
 if ($2 > $3)  exit 1
 if ($4 < 1)   exit 1
-if ($5 <= 0)   exit 1
+if ($5 <= 0)  exit 1
 
 
 set tmp = `mktemp`
@@ -33,7 +34,6 @@ if ($?) exit 1
 setIntersect.sh $tmp.start-new $tmp.final 0 > $tmp.start-good
 if ($?) exit 1
 
-
 set N = `wc -l $tmp.start-good`
 echo "Tree size: $N[1]"
 @ S = $N[1] / 2
@@ -41,22 +41,32 @@ set S_max = 300  # PAR
 if ($S > $S_max)  set S = $S_max
 echo "Sample size: $S"
 
-setRandOrd $tmp.start-good $5 | head -$S | sort > $tmp.test
-
-
-echo ""
-echo "Complete optimization  ..."
-list2pairs $tmp.test > $tmp.pairs
-if ($?) exit 1
-
-distTree_inc_request2dissim.sh $1 $tmp.pairs $tmp.dissim
-if ($?) exit 1
-
-pairs2attr2 $tmp.dissim 1 cons 6 -distance > $tmp.dm
-if ($?) exit 1
-
-makeDistTree  -data $tmp  -dissim cons  -optimize  -output_tree $tmp.tree  -output_feature_tree $tmp.feature_tree1 | grep "# Discernible leaves:"
-if ($?) exit 1
+setRandOrd $tmp.start-good $5 | head -$S > $tmp.test
+if (-z $tmp.test) then
+	cp $tmp.init $tmp.test
+	if ($?) exit 1
+  cp $1/hist/tree.$2 $tmp.tree
+	if ($?) exit 1
+	makeDistTree  -input_tree $tmp.tree  -output_feature_tree $tmp.feature_tree1 | grep "# Discernible leaves:"
+	if ($?) exit 1
+else
+	echo ""
+	echo "Complete optimization  ..."
+	cat $tmp.init >> $tmp.test
+	sort.sh $tmp.test
+	
+	list2pairs $tmp.test > $tmp.pairs
+	if ($?) exit 1
+	
+	distTree_inc_request2dissim.sh $1 $tmp.pairs $tmp.dissim
+	if ($?) exit 1
+	
+	pairs2attr2 $tmp.dissim 1 cons 6 -distance > $tmp.dm
+	if ($?) exit 1
+	
+	makeDistTree  -data $tmp  -dissim cons  -optimize  -output_tree $tmp.tree  -output_feature_tree $tmp.feature_tree1 | grep "# Discernible leaves:"
+	if ($?) exit 1
+endif
 
 makeFeatureTree  -input_tree $tmp.feature_tree1  -features $1/phen  -output_core $tmp.core  -qual $tmp.qual  
 if ($?) exit 1
@@ -73,5 +83,7 @@ while ($i <= $3)
 	@ n = $n + 1
 end
 
+distTree_inc_quality_phen_step.sh $1 $tmp.test opt $tmp
+if ($?) exit 1
 
 rm -fr $tmp.*
